+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.1.1] - 2026-03-23
+
+### Fixed
+
+* `getAll()` crash (`ClassCastException`) when preferences contain `StringSet` values
+
+### Changed
+
+* Sample app rewritten to showcase all library features (all data types, getAll, async callbacks, change listeners, remove, clear, close/reinitialize)
+
 ## [5.1.0] - 2026-03-23
 
 ### Added